SBERDOMA-299 review fixes
