+++ apps/condo/domains/ticket/schema/TicketFile.js
@@ -13,6 +13,7 @@ const FileAdapter = require('@condo/domains/common/utils/fileAdapter')
 const Adapter = new FileAdapter('ticket')
 
 // zuch(todo): find a way to upload images in jest tests
+// and make file field required
 const TicketFile = new GQLListSchema('TicketFile', {
     schemaDoc: 'File attached to the ticket',
     fields: {
